+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+-   [Patch] Call `onOpenFinish` when the modal is immediately open as it is mounted.
+
 ## 0.6.0 - 2019-05-01
 
 ### Added